Move unit tests to integration tests
